Task 1 #2

+++ ex01.hpp
@@ -2,7 +2,7 @@
 //change the below code
 // the function below now only works with integers, change it to work with any type
 // also fill out the body
-template<typename T = long double>
+template<typename T = int>
 void multiples (T& sum, T x, int n)
 {
     for (int i = 1; i <= n; ++i) {
